test_hc

+++ dalgrak/src/main/java/springbootApplication/repository/TimerRepository.java
@@ -1,7 +1,8 @@
 package springbootApplication.repository;
 
-import springbootApplication.domain.Timer; // 올바른 경로로 수정
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TimerRepository extends JpaRepository<Timer, Long> {
+import springbootApplication.domain.Timer;
+
+public interface TimerRepository extends JpaRepository<Timer, Integer> {
 }
